rays: Move render_buffer and frame_buffer to opengl/
Both are entirely OpenGL-specific. Update include paths in
texture.cpp, painter.cpp, and platform bitmap files.
