+++ rays/src/opengl/render_buffer.cpp
