+++ rays/src/win32/bitmap.cpp
@@ -10,7 +10,7 @@
 #include "../opengl/color_space.h"
 #include "../font.h"
 #include "../texture.h"
-#include "../frame_buffer.h"
+#include "../opengl/frame_buffer.h"
 #include "gdi.h"
